chore: consolidate util helpers
