+++ api/src/app/util/__init__.py
@@ -1,5 +1,2 @@
-"""Utility helpers for assembling snapshot views."""
-
-# Re-export helpers from legacy plural package so old imports keep working
-from ..utils.auth_helpers import verify_jwt            # type: ignore
-from ..utils.supabase_client import supabase_client    # type: ignore
+"""❗ DEPRECATED – use app.utils instead. Will be removed next release."""
+from ..utils import *  # type: ignore  # re-export everything temporarily  # noqa: F403
