+++ api/src/app/utils/jwt.py
@@ -1,0 +1,5 @@
+"""TEMP stub JWT helpers – replace with real verification later."""
+
+def verify_jwt(_: str) -> dict[str, str]:
+    # ↵️  TODO: real JWT check
+    return {"user_id": "stub-user", "workspace_id": "stub-ws"}
